requirements.txt
Flask>=2.3.0
pandas>=2.0.0
numpy>=1.25.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
xlrd>=2.0.0
xgboost>=1.7.0
lightgbm>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0
blinker>=1.6.0
requests>=2.31.0
python-dotenv>=1.0.0 